Remove data-uri encoding of woff font.
In researching WordPress' Twentyfifteen theme performance, I found that
data-uri enconding the woff Genericons font file could cause visually
complete delays of up to 300ms on a mobile connection.

When a browser tries to render a page, it tries to complete the CSSOM as
quickly as possible so the first paint can be completed. The CSSOM is
dependent upon the CSS being downloaded from the server(s). Any delays
in getting the CSS can cause delays in rendering the page.

The Genericons CSS files weighs in around 16k and can be reduced to 2k
with the encoded woff file removed. The size reduction is significant
and can lead to much quicker resolution of the CSSOM if this change is
made.

Furthermore, data-uri encoding the font into the CSS requires that all
browsers receive the font even if it cannot be used by the browser. If
all fonts are located at external locations, the browser can make the
decision about which ones to load instead of the application, which is
not in a good place to make such a decision.

This commit removes the data-uri encoded font and prefers the external
font file instead.