Implement basic spriting of SVGs.
This is using the <symbol> method of insertion—there are a few, but this one is my favourite.
To provide maximum user-flexibility, I've set it up so that the SVG can be called externally (default, caches better)
or internally injected into the page (better for IE compatibility and possibly multi-colour icons). The external
method is default, but this can be overridden by passing an extra parameter to the get_genericon() function.

I've also added a simple-simple genericon() function that echoes out the genericon, to save keystrokes.

In case users don't want to work with sprites for some reason, but still want SVG (maybe they're doing complex
animation stuff or multi-coloured icons or whatnot, or need to support super old browsers), I've left the original
script in the folder, but we may want to remove it if this turns out not to be super useful.